+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/media/RichMessageMemberTelegramMedia.kt
@@ -1,0 +1,7 @@
+package dev.inmo.tgbotapi.types.media
+
+import kotlinx.serialization.Serializable
+
+@Suppress("SERIALIZER_TYPE_INCOMPATIBLE")
+@Serializable(TelegramMediaSerializer::class)
+sealed interface RichMessageMemberTelegramMedia : TelegramMedia